refactor: move `TimeoutStackStyle` to its own file (#6216)
