+++ src/common/TimeoutStackStyle.hpp
@@ -1,0 +1,11 @@
+namespace chatterino {
+
+enum class TimeoutStackStyle : int {
+    StackHard = 0,
+    DontStackBeyondUserMessage = 1,
+    DontStack = 2,
+
+    Default = DontStackBeyondUserMessage,
+};
+
+}  // namespace chatterino
