+++ src/singletons/Settings.hpp
@@ -1,10 +1,10 @@
 #pragma once
 
-#include "common/Channel.hpp"
 #include "common/ChatterinoSetting.hpp"
 #include "common/enums/MessageOverflow.hpp"
 #include "common/Modes.hpp"
 #include "common/SignalVector.hpp"
+#include "common/TimeoutStackStyle.hpp"
 #include "controllers/filters/FilterRecord.hpp"
 #include "controllers/highlights/HighlightBadge.hpp"
 #include "controllers/highlights/HighlightBlacklistUser.hpp"
